etl-pipeline/requirements.txt
alembic==1.10.2
beautifulsoup4==4.12.0
boto3==1.26.104
elastic-transport==8.4.0
elasticsearch-dsl==7.4.1
elasticsearch==7.16.3
fasttext==0.9.2
flasgger==0.9.5
flask==2.2.4
flask-cors==5.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
lxml==4.9.2
numpy==1.24.2
pandas==2.0.0
pika==1.3.1
pillow==9.5.0
psycopg2-binary==2.9.6
pypdf==5.4.0
python-levenshtein==0.20.4
pycountry==22.3.5
pyjwt[crypto]==2.6.0
pymarc==4.2.2
pyyaml==6.0
redis==4.5.4
requests==2.28.2
requests_oauthlib==1.3.1
requests-aws4auth==1.3.1
scikit-learn==1.2.2
sqlalchemy==2.0.20
waitress==3.0.1
werkzeug==2.2.2
newrelic==8.8.0
geocoder==1.38.1
